feat: add utils
